Refactor identity guide modalize
